--- kag/open_benchmark/hotpotqa/benchmark.txt
@@ -1,2 +1,0 @@
-
-processNum: 1	em: 0.0	f1: 0.0	answer_similarity: 0.0	LLM-Accuracy: 0.0	cost: 18.251737117767334	task_name: hotpotqa
