+++ kag/open_benchmark/musique/.gitignore
@@ -1,3 +1,3 @@
 ckpt/
-/solver/musique_res_*.json
-/solver/musique_metrics_*.json
+/solver/2wiki_res_*.json
+/solver/2wiki_metrics_*.json
