+++ kag/open_benchmark/musique/README_cn.md
@@ -1,11 +1,11 @@
-# KAG 示例：MuSiQue
+# KAG 示例：HotpotQA
 
 [English](./README.md) |
 [简体中文](./README_cn.md)
 
-[MuSiQue](https://arxiv.org/abs/2108.00573) 是一个用于对推理步骤进行全面评估的多跳问答数据集。[KAG](https://arxiv.org/abs/2409.13731) 和 [HippoRAG](https://arxiv.org/abs/2405.14831) 用它评估多跳问答的性能。
+[HotpotQA](https://arxiv.org/abs/1809.09600) 是一个用于多样和可解释多跳问答的数据集。[KAG](https://arxiv.org/abs/2409.13731) 和 [HippoRAG](https://arxiv.org/abs/2405.14831) 用它评估多跳问答的性能。
 
-本例我们展示为 MuSiQue 数据集构建知识图谱，然后用 KAG 为评估问题生成答案，并与标准答案对比计算 EM 和 F1 指标。
+本例我们展示为 HotpotQA 数据集构建知识图谱，然后用 KAG 为评估问题生成答案，并与标准答案对比计算 EM 和 F1 指标。
 
 ## 1. 前置条件
 
@@ -16,7 +16,7 @@
 ### Step 1：进入示例目录
 
 ```bash
-cd kag/open_benchmark/musique
+cd kag/open_benchmark/hotpotqa
 ```
 
 ### Step 2：配置模型
@@ -35,7 +35,7 @@ knext project restore --host_addr http://127.0.0.1:8887 --proj_path .
 
 ### Step 4：提交 schema
 
-执行以下命令提交 schema [MuSiQue.schema](./schema/MuSiQue.schema)。
+执行以下命令提交 schema [HotpotQA.schema](./schema/HotpotQA.schema)。
 
 ```bash
 knext schema commit
@@ -57,9 +57,9 @@ cd builder && python indexer.py && cd ..
 cd solver && python eva.py --qa_file ./data/qa_sub.json && cd ..
 ```
 
-生成的答案被保存至 ``./solver/musique_res_*.json``.
+生成的答案被保存至 ``./solver/hotpotqa_res_*.json``.
 
-计算出的 EM 和 F1 指标被保存至 ``./solver/musique_metrics_*.json``.
+计算出的 EM 和 F1 指标被保存至 ``./solver/hotpotqa_metrics_*.json``.
 
 ### Step 7：（可选）清理
