+++ kag/open_benchmark/musique/kag_config_chunk.yaml
@@ -1,24 +1,12 @@
-openie_llm: &openie_llm
+llm: &llm
   type: maas
   base_url: https://dashscope.aliyuncs.com/compatible-mode/v1/
-  api_key: sk-
+  api_key: !ENV LLM_API_KEY
   model: qwen2.5-72b-instruct
 
-chat_llm: &chat_llm
-  type: maas
-  base_url: https://dashscope.aliyuncs.com/compatible-mode/v1/
-  api_key: sk-
-  model: qwen2.5-72b-instruct
-  stream: false
-
-ner_llm:
-  type: maas
-  base_url: https://dashscope.aliyuncs.com/compatible-mode/v1/
-  api_key: sk-
-  model: qwen2.5-72b-instruct
 
 vectorize_model: &vectorize_model
-  api_key: sk-
+  api_key: !ENV VECTORIZE_MODEL_API_KEY
   base_url: https://api.siliconflow.cn/v1/
   model: BAAI/bge-m3
   type: openai
@@ -63,7 +51,7 @@ kag_builder_pipeline:
   num_threads_per_chain: 1
   num_chains: 16
   scanner:
-    type: 2wiki_dataset_scanner # kag.builder.component.scanner.dataset_scanner.MusiqueCorpusScanner
+    type: musique_dataset_scanner # kag.builder.component.scanner.dataset_scanner.MusiqueCorpusScanner
 #------------kag-builder configuration end----------------#
 
 
@@ -85,7 +73,7 @@ r1: &r1
 
 retriever_executor: &retriever
   type: evidence_based_reasoner
-  llm: *chat_llm
+  llm: *llm
   retrievers:
     - *r1
   merger:
@@ -97,7 +85,7 @@ kag_solver_pipeline:
   type: kag_static_pipeline
   planner:
     type: kag_static_planner
-    llm: *chat_llm
+    llm: *llm
     plan_prompt:
       type: retriever_static_planning
     rewrite_prompt:
@@ -106,5 +94,5 @@ kag_solver_pipeline:
     - *retriever
   generator:
     type: llm_generator_with_thought
-    llm_client: *chat_llm
+    llm_client: *llm
 #------------kag-solver configuration end----------------#
