+++ kag/open_benchmark/musique/kag_config_simple.yaml
@@ -21,13 +21,22 @@ project:
   host_addr: http://127.0.0.1:8887
   id: '1'
   language: en
-  namespace: HotpotQA
+  namespace: MuSiQue
 #------------project configuration end----------------#
 
 #------------kag-builder configuration start----------------#
 e1: &e1
   type: naive_rag_extractor
 
+e2: &e2
+  type: schema_free_extractor # kag.builder.component.extractor.schema_free_extractor.SchemaFreeExtractor
+  llm: *llm
+  ner_prompt:
+    type: default_ner # kag.builder.prompt.default.ner.OpenIENERPrompt
+  std_prompt:
+    type: default_std # kag.builder.prompt.default.std.OpenIEEntitystandardizationdPrompt
+  triple_prompt:
+    type: default_triple # kag.builder.prompt.default.triple.OpenIETriplePrompt
 
 
 kag_builder_pipeline:
@@ -51,7 +60,7 @@ kag_builder_pipeline:
   num_threads_per_chain: 1
   num_chains: 16
   scanner:
-    type: hotpotqa_dataset_scanner # kag.builder.component.scanner.dataset_scanner.MusiqueCorpusScanner
+    type: musique_dataset_scanner # kag.builder.component.scanner.dataset_scanner.MusiqueCorpusScanner
 #------------kag-builder configuration end----------------#
 
 
@@ -64,8 +73,7 @@ search_api: &search_api
 graph_api: &graph_api
   type: openspg_graph_api #kag.solver.tools.graph_api.impl.openspg_graph_api.OpenSPGGraphApi
 
-
-r1: &r1
+r2: &r2
   type: vector_chunk_retriever
   vectorize_model: *vectorize_model
   search_api: *search_api
@@ -75,7 +83,7 @@ retriever_executor: &retriever
   type: evidence_based_reasoner
   llm: *llm
   retrievers:
-    - *r1
+    - *r2
   merger:
     type: kag_merger
