+++ kag/open_benchmark/musique/kag_config_summary.yaml
@@ -1,12 +1,24 @@
-llm: &llm
+openie_llm: &openie_llm
   type: maas
   base_url: https://dashscope.aliyuncs.com/compatible-mode/v1/
-  api_key: !ENV LLM_API_KEY
+  api_key: sk-
   model: qwen2.5-72b-instruct
 
+chat_llm: &chat_llm
+  type: maas
+  base_url: https://dashscope.aliyuncs.com/compatible-mode/v1/
+  api_key: sk-
+  model: qwen2.5-72b-instruct
+  stream: false
+
+ner_llm:
+  type: maas
+  base_url: https://dashscope.aliyuncs.com/compatible-mode/v1/
+  api_key: sk-
+  model: qwen2.5-72b-instruct
 
 vectorize_model: &vectorize_model
-  api_key: !ENV VECTORIZE_MODEL_API_KEY
+  api_key: sk-
   base_url: https://api.siliconflow.cn/v1/
   model: BAAI/bge-m3
   type: openai
@@ -26,10 +38,10 @@ project:
 
 #------------kag-builder configuration start----------------#
 e1: &e1
-  type: atomic_query_extractor
-  llm: *llm
-  prompt:
-    type: atomic_query_extract
+  type: chunk_extractor
+
+e2: &e2
+  type: summary_extractor
 
 
 kag_builder_pipeline:
@@ -37,6 +49,7 @@ kag_builder_pipeline:
     type: unstructured_builder_chain # kag.builder.default_chain.DefaultUnstructuredBuilderChain
     extractor:
       - *e1
+      - *e2
     reader:
       type: dict_reader # kag.builder.component.reader.dict_reader.DictReader
     post_processor:
@@ -66,22 +79,23 @@ search_api: &search_api
 graph_api: &graph_api
   type: openspg_graph_api #kag.solver.tools.graph_api.impl.openspg_graph_api.OpenSPGGraphApi
 
-r1: &r1
-  type: atomic_query_chunk_retriever
-  vectorize_model: *vectorize_model
-  search_api: *search_api
-  graph_api: *graph_api
-  top_k: 10
 
-r2: &r2
+r1: &r1
   type: vector_chunk_retriever
   vectorize_model: *vectorize_model
   search_api: *search_api
   top_k: 10
 
+r2: &r2
+  type: summary_chunk_retriever
+  search_api: *search_api
+  graph_api: *graph_api
+  vectorize_model: *vectorize_model
+  top_k: 5
+
 retriever_executor: &retriever
   type: evidence_based_reasoner
-  llm: *llm
+  llm: *chat_llm
   retrievers:
     - *r1
     - *r2
@@ -94,7 +108,7 @@ kag_solver_pipeline:
   type: kag_static_pipeline
   planner:
     type: kag_static_planner
-    llm: *llm
+    llm: *chat_llm
     plan_prompt:
       type: retriever_static_planning
     rewrite_prompt:
@@ -103,5 +117,5 @@ kag_solver_pipeline:
     - *retriever
   generator:
     type: llm_generator_with_thought
-    llm_client: *llm
+    llm_client: *chat_llm
 #------------kag-solver configuration end----------------#
