--- kag/open_benchmark/musique/reasoner/__init__.py
@@ -1,20 +1,0 @@
-# Copyright 2023 OpenSPG Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License. You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the License
-# is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-# or implied.
-
-"""
-Place the DSL file for graph reasoning in this directory.
-For example:
-
-```company.dsl
-MATCH (s:DEFAULT.Company)
-RETURN s.id, s.address
-```
-"""
