+++ kag/open_benchmark/musique/src/llm_genereator_with_thought.py
